--- apps/dash-turbine-maintenance/.gitignore
@@ -1,5 +1,0 @@
-.idea/*
-__pycache__/
-*.ipynb_checkpoints/
-
-
