+++ apps/dash-turbine-maintenance/assets/images/dash-logo.png
