+++ apps/dash-turbine-maintenance/constants.py
@@ -1,0 +1,1 @@
+gauge_size = "auto"
