+++ apps/dash-turbine-maintenance/requirements.txt
@@ -1,15 +1,11 @@
-dash==1.19.0
-dash-bootstrap-components==0.11.3
-dash-core-components==1.15.0
+dash==2.4.1
+pandas==1.4.2
+gunicorn==20.1.0
+dash-bootstrap-components==1.1.0
 dash-daq==0.5.0
-dash-html-components==1.1.2
-dash-renderer==1.9.0
-dash-table==4.11.2
+dash-renderer==1.9.1
 matplotlib==3.3.4
-numpy==1.19.5
-pandas==1.1.5
-plotly==4.14.3
 scikit-learn==0.24.1
 scipy==1.5.4
 xgboost==1.3.3
-gunicorn
+
